Use validation logic on every yaml description to avoid undefined behaviour
