+++ binaries/daemon/src/lib.rs
@@ -26,10 +26,7 @@ use std::{
 };
 use tcp_utils::tcp_receive;
 use tokio::sync::mpsc::UnboundedSender;
-use tokio::{
-    fs,
-    sync::{mpsc, oneshot},
-};
+use tokio::sync::{mpsc, oneshot};
 use tokio_stream::{wrappers::ReceiverStream, Stream, StreamExt};
 use uuid::Uuid;
